Refactor music to its own files.
