+++ music.h
@@ -1,0 +1,5 @@
+void run_voices(void);
+void init_voices(void);
+
+void new_level_music(void);
+void run_music(void);
